+++ lib/avs_image.dart
@@ -81,9 +81,17 @@ class AVSImage extends StatelessWidget {
   //Image zoom start style. Default onTap
   final ZoomStyle zoomStyle;
 
-  /// Image On Tap. if this is not null and zoomStyle.onTap then zoom is false
+  /// Image onTap function. if this is not null and zoomStyle.onTap then zoom is false
   ///
   final Function()? onTap;
+
+  /// Image onLongPress function. if this is not null and zoomStyle.onLongPress then zoom is false
+  ///
+  final Function()? onLongPress;
+
+  /// Image onDoubleTap function. if this is not null and zoomStyle.onDoubleTap then zoom is false
+  ///
+  final Function()? onDoubleTap;
 
   /// Loading Progress Indicator
   ///
@@ -112,6 +120,8 @@ class AVSImage extends StatelessWidget {
     this.cachedImage = true,
     this.showProgressIndicator = true,
     this.onTap,
+    this.onDoubleTap,
+    this.onLongPress,
     this.progressIndicatorWidget,
   })  : isSvg = _isSvgCheck(path),
         isLocalPosition = _isLocalImageCheck(path),
@@ -128,9 +138,7 @@ class AVSImage extends StatelessWidget {
           child: AspectRatio(
             aspectRatio: 1,
             child: ClipRRect(
-              borderRadius: isCircle == true
-                  ? BorderRadius.circular(360)
-                  : radius ?? BorderRadius.zero,
+              borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
               child: GestureDetector(
                 onTap: () async => _onTapFunction(context),
                 onLongPress: () async => _onLongPressFunction(context),
@@ -142,9 +150,7 @@ class AVSImage extends StatelessWidget {
         );
       } else {
         return ClipRRect(
-          borderRadius: isCircle == true
-              ? BorderRadius.circular(360)
-              : radius ?? BorderRadius.zero,
+          borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
           child: GestureDetector(
             onTap: () async => _onTapFunction(context),
             onLongPress: () async => _onLongPressFunction(context),
@@ -155,9 +161,7 @@ class AVSImage extends StatelessWidget {
       }
     } else {
       return ClipRRect(
-        borderRadius: isCircle == true
-            ? BorderRadius.circular(360)
-            : radius ?? BorderRadius.zero,
+        borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
         child: GestureDetector(
           onTap: () async => _onTapFunction(context),
           onLongPress: () async => _onLongPressFunction(context),
@@ -183,23 +187,18 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildLocalSVG() {
-    return gradient == null
-        ? _buildLocalSVGNoGradient()
-        : _buildLocalSVGWithGradient();
+    return gradient == null ? _buildLocalSVGNoGradient() : _buildLocalSVGWithGradient();
   }
 
   Widget _buildLocalSVGNoGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: SvgPicture.asset(
         url,
         height: height,
         width: width,
         fit: _defaultFit,
-        colorFilter:
-            color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
+        colorFilter: color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
         alignment: alignment,
       ),
     );
@@ -207,12 +206,9 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildLocalSVGWithGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: _SVGLinearGradientMask(
-        gradient: gradient ??
-            const LinearGradient(colors: [Colors.black, Colors.black26]),
+        gradient: gradient ?? const LinearGradient(colors: [Colors.black, Colors.black26]),
         child: SvgPicture.asset(
           url,
           height: height,
@@ -226,23 +222,18 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildNetworkSVG() {
-    return gradient == null
-        ? _buildNetworkSVGNoGradient()
-        : _buildNetworkSVGWithGradient();
+    return gradient == null ? _buildNetworkSVGNoGradient() : _buildNetworkSVGWithGradient();
   }
 
   Widget _buildNetworkSVGNoGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: SvgPicture.network(
         url,
         height: height,
         width: width,
         fit: _defaultFit,
-        colorFilter:
-            color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
+        colorFilter: color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
         alignment: alignment,
       ),
     );
@@ -250,12 +241,9 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildNetworkSVGWithGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: _SVGLinearGradientMask(
-        gradient: gradient ??
-            const LinearGradient(colors: [Colors.black, Colors.black26]),
+        gradient: gradient ?? const LinearGradient(colors: [Colors.black, Colors.black26]),
         child: SvgPicture.network(
           url,
           height: height,
@@ -270,9 +258,7 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildLocalImage() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: Image.asset(
         url,
         color: color,
@@ -289,16 +275,12 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildNetworkImage() {
-    return cachedImage == true
-        ? _buildNetworkImageWithCache()
-        : _buildNetworkImageNoCache();
+    return cachedImage == true ? _buildNetworkImageWithCache() : _buildNetworkImageNoCache();
   }
 
   Widget _buildNetworkImageNoCache() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: Image.network(
         url,
         color: color,
@@ -312,10 +294,7 @@ class AVSImage extends StatelessWidget {
           } else {
             return showProgressIndicator
                 ? _buildProgressIndicator(
-                    ((loadingProgress.cumulativeBytesLoaded) /
-                                (loadingProgress.expectedTotalBytes ?? 1))
-                            .toDouble() *
-                        100.toDouble(),
+                    ((loadingProgress.cumulativeBytesLoaded) / (loadingProgress.expectedTotalBytes ?? 1)).toDouble() * 100.toDouble(),
                   )
                 : const SizedBox();
           }
@@ -330,9 +309,7 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildNetworkImageWithCache() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: CachedNetworkImage(
         imageUrl: url,
         height: height,
@@ -341,10 +318,7 @@ class AVSImage extends StatelessWidget {
         fit: _defaultFit,
         maxHeightDiskCache: 3000,
         color: color,
-        progressIndicatorBuilder: (context, url, progress) =>
-            showProgressIndicator
-                ? _buildProgressIndicator(progress.downloaded.toDouble())
-                : const SizedBox(),
+        progressIndicatorBuilder: (context, url, progress) => showProgressIndicator ? _buildProgressIndicator(progress.downloaded.toDouble()) : const SizedBox(),
         errorWidget: (context, url, error) {
           _showLog("Network Image With Cache Exploded: $url");
 
@@ -392,7 +366,9 @@ class AVSImage extends StatelessWidget {
   }
 
   Future<void>? _onLongPressFunction(BuildContext context) async {
-    if (zoomStyle == ZoomStyle.onLongPress && zoom == true) {
+    if (onLongPress != null) {
+      onLongPress!();
+    } else if (zoomStyle == ZoomStyle.onLongPress && zoom == true) {
       _zoomFunc(context, url);
     } else {
       return;
@@ -400,7 +376,9 @@ class AVSImage extends StatelessWidget {
   }
 
   Future<void>? onDoubleTapFunction(BuildContext context) async {
-    if (zoomStyle == ZoomStyle.onDoubleTap && zoom == true) {
+    if (onDoubleTap != null) {
+      onDoubleTap!();
+    } else if (zoomStyle == ZoomStyle.onDoubleTap && zoom == true) {
       _zoomFunc(context, url);
     } else {
       return;
@@ -421,12 +399,7 @@ ImageProvider AVSImageProvider(
   bool isLocalPosition = _isLocalImageCheck(path);
 
   if (isSvg) {
-    return AVSSVGProvider(path,
-        color: color,
-        scale: scale,
-        height: height,
-        width: width,
-        gradient: gradient);
+    return AVSSVGProvider(path, color: color, scale: scale, height: height, width: width, gradient: gradient);
   } else if (isSvg == false && isLocalPosition) {
     return AssetImage(path);
   } else if (isSvg == false && !isLocalPosition) {